YAML: move getBody/withBody logic to a BlockScalar trait
The prior promotion of getBody/withBody onto Yaml.Scalar itself breaks
under any Moderne CLI whose bundled rewrite-yaml predates the API
addition: the CLI's parent classloader wins on org.openrewrite.yaml.tree.Yaml$Scalar,
so recipe artifacts compiled against the new instance methods hit
NoSuchMethodError when scalars arrive without them. Verified with mod
4.3.7 + our SNAPSHOT installed as a recipe jar.

Move the same logic into a new BlockScalar trait in
org.openrewrite.yaml.trait, alongside YamlValue / YamlReference. The
trait class ships in the recipe jar; its methods only depend on
Yaml.Scalar's already-stable API (getValue / getStyle / withValue), so
customer recipes that adopt the new API keep working under older CLI
bundles.

Consumers refactored: ChangePropertyValue, ChangeValue, YamlValue,
YamlReference. Yaml.Scalar keeps only a Javadoc note pointing at the
trait.

Follow-up #8157's MergeYamlVisitor will pick up the trait in its own
rebase.
