+++ rewrite-yaml/src/main/java/org/openrewrite/yaml/ChangePropertyValue.java
@@ -22,6 +22,7 @@ import org.openrewrite.*;
 import org.openrewrite.internal.ListUtils;
 import org.openrewrite.internal.NameCaseConvention;
 import org.openrewrite.internal.StringUtils;
+import org.openrewrite.yaml.trait.BlockScalar;
 import org.openrewrite.yaml.tree.Yaml;
 
 import java.util.Iterator;
@@ -96,7 +97,7 @@ public class ChangePropertyValue extends Recipe {
                 Yaml.Mapping.Entry e = super.visitMappingEntry(entry, ctx);
                 String prop = getProperty(getCursor());
                 if (keyMatcher.matchesGlob(prop) && matchesOldValue(e.getValue())) {
-                    Yaml.Block updatedValue = updateValue(e.getValue());
+                    Yaml.Block updatedValue = updateValue(e.getValue(), getCursor());
                     if (updatedValue != null) {
                         e = e.withValue(updatedValue);
                     }
@@ -107,23 +108,23 @@ public class ChangePropertyValue extends Recipe {
     }
 
     // returns null if value should not change
-    private Yaml.@Nullable Block updateValue(Yaml.Block value) {
+    private Yaml.@Nullable Block updateValue(Yaml.Block value, Cursor parent) {
         if (value instanceof Yaml.Scalar) {
-            Yaml.Scalar scalar = (Yaml.Scalar) value;
-            String body = scalar.getBody();
+            BlockScalar block = BlockScalar.of(new Cursor(parent, value));
+            String body = block.getBody();
             String updatedBody = Boolean.TRUE.equals(regex) ?
                     body.replaceAll(Objects.requireNonNull(oldValue), newValue) :
                     newValue;
             if (body.equals(updatedBody)) {
                 return null;
             }
-            return scalar.withBody(updatedBody);
+            return block.withBody(updatedBody);
         }
         if (value instanceof Yaml.Sequence) {
             Yaml.Sequence sequence = (Yaml.Sequence) value;
             return sequence.withEntries(ListUtils.map(sequence.getEntries(), entry -> {
                 if (matchesOldValue(entry.getBlock())) {
-                    Yaml.Block updatedValue = updateValue(entry.getBlock());
+                    Yaml.Block updatedValue = updateValue(entry.getBlock(), parent);
                     if (updatedValue != null) {
                         return entry.withBlock(updatedValue);
                     }
